Dirty hack to fix () wrapping expressions
Fix involves copying the lexer. On seeing '(', parser copies the
lexer and attempts to parse '(<expr>)'. If this succeeds, repeats
the parse on the original lexer. If this fails, it attempts to
parse a tuple.

This has a few drawbacks. Nested tuples/expressions get parsed
O(2^depth) times and the reported error is unhelpful for invalid
expressions (thinks the string was intended to be a tuple).