fixes #25 multiplication in an arrow operation
writing `. *` in an arrow operation (>>, ->, etc) are parsed as a get operation
instead of a math operation. While it is allowed in the grammar to do a get
operation with an asterisk, it is unclear as to what it does as it is not
handled yet during parsing. Removing the asterisk from `get` grammar is a very
fast fix for this problem.