Fix #356
Simplify creating array from set to account for holes, which were not accounted for.

New algorithm needs to loop through the set twice, once to get the index bounds of its elements, then again to populate the []Value array.

Also fixed an incidental panic when trying to stringify an array with holes.