Improve macro impl, merge syntax with grammar
Macros are becoming a generic way to embed code in non-arrai synax within arrai, alongside a grammar with which to parse it. If a transform is provided with the grammar, the parsed AST is transformed in the parsing processed. If no transform is provided, the existing grammar behaviour is largely retained.