Implement macro support in Arrai parsing (#453)
Implement syntax for macros: grammars with transforms

Fixes #440 

Replace inline grammar embedding term ({:grammar:content:}) with upgraded macro term ({:macro[rule]?:content:}). macro can be a plain grammar AST, or a tuple containing @grammar and @transform keys. @transform contains functions that are applied to the content AST produced by parsing it with @grammar. As before, the output of the macro effectively replaces the {:...:} fragment of the program.

If a rule is not explicitly provided, assume a default of the first rule in the grammar. If that fails to parse, panic.

Similarly, if @transform is explicitly provided but doesn't contain a function corresponding to the name of the rule used to parse, panic.

Future work will progress the macro syntax (from {:macro[rule]:content:} to macro[rule]{content}).