Optimize nest for single attribute (#559)
The form `x nest |y|y` would nest all values that correspond to the `y`
attribute in a set of tuples with one attribute. This PR introduces
a shorter form `x nest y` that does the same thing except, it is nested
into a set of values instead of tuples with one attribute for
optimization.