allow nested import root and import root fixes
This commit will allow nested root markers in a project.
Root markers will be found by going up the source directory
of a file. A root marker can be in the same directory or the parent
directories. The root marker locations will also be cached for faster
root marker lookup.

This commit also fixed the root marker lookup as initially it was
done by searching from the provided import path instead of the directory
location of the source scripts.