Disable HTML escaping in JSON encoder
By default, json.Marshal escapes special chars such as '>' to make it safe for HTML. This is not the typical use case for arrai's JSON exporter.

Details here: https://golang.org/pkg/encoding/json/#Marshal

In addition, the change also adds a newline character to the end of the output, which is desirable in most cases.