fixes #117, fix multiple fallback patterns in tuple (#118)
Initially, multiple fallback patterns would trigger non-deterministic pattern
error.

```
@> let (?:a:1, ?:b:2) = (b: 2); a
2021-01-27T13:57:49.186686+11:00 error_message=non-deterministic pattern is not supported yet

.:1:1:
let (?:a:1, ?:b:2) = (b: 2); a INFO
```

This is also triggered when fallback pattern and extra patterns are used

```
@> let (?:a:1, ...) = (b: 2); a
2021-01-27T13:58:40.271975+11:00 error_message=non-deterministic pattern is not supported yet

.:1:1:
let (?:a:1, ...) = (b: 2); a INFO
```

This commit fixes that. The implementation also fixes several issues:

1. Making sure that `...` catches whatever is left after matching the rest of the patterns.

Before:

```
@> let (a: 1, ...tail, c: 3) = (a: 1, b: 2, c: 3); tail
(b: 2, c: 3)
```

Now:

```
@> let (a: 1, ...tail, c: 3) = (a: 1, b: 2, c: 3); tail
(b: 2)
```

2. Fallback pattern behavior. Initially, fallback pattern does not fail even when
there are missing elements.

```
@> let (?:a:1) = (b: 2); a
1
```

That expression should match `a` and nothing else should be left on the tuple on the right
hand side. This commit fixes that.

Now:

```
@> let (?:a:1) = (b: 2); a
2021-01-27T14:03:07.897535+11:00 error_message=length of tuple (b: 2) longer than tuple pattern (a?: a:1)

.:1:1:
let (?:a:1) = (b: 2); a INFO
```