Move shell from internal to pkg
`Shell` is a very useful feature for debugging issues when runtime errors are thrown. It's already used by main.go to catch errors in processes started from the arr.ai CLI, but since the arr.ai runtime is also a library, there are other contexts in which it could be useful.

Exporting the `Shell` function allows Go programs that use arr.ai (e.g. `EvaluateExpr`) to benefit from the same debugging tools.