import cache resets everytime EvaluateExpr is called (#212)
cache was implemented as a global variable and when arrai code is called from golang,
the import cache persists. This commit moves import cache to context and ensures it
is reset every time arrai code is called in a golang code. 

It also allows `EvaluateExpr` to receive a context argument with existing cache inside. The function will not override this cache so that cache can be reused. This is meant for those who run arrai scripts in a golang program.