Support non-strict JSON encode/decode (#218)
Add //encoding.json.encoder and //encoding.json.decoder. These functions take a config tuple, and return configured encode and decode functions respectively.

In particular, JSON decoding and encoding can be configured with the strict flag. (strict: true) is today's behaviour, wherein JSON-compatible types that can be empty (strings, bools, dictionaries) must be wrapped in tuples ((s: "") etc.). With (strict: false), those tuples are not used, and empty values map to null in both directions. This is often preferable if the original type of empty values is not useful.

This feature obsoletes the need for the contrib function toJson, which can be very expensive for large inputs.

The encoder config offers more options, including prefix and indent for controlling formatted output, and escapeHTML to replace characters that may cause trouble in HTML contexts (<, > and &).

Other changes:

- Support //encoding.yaml.decoder(()), although decoding doesn't accept any config attrs.
- Change strict encoding to error if a tuple contains a type discriminator plus anything else (e.g. (a: [], s: ""))