fix issue with mutability of Relation in compose operations (#247)
Previously, a compose operation would modify the LHS set after evaluating. This is due to the use slices which results in pointers. When slices are appended, the original array that was sliced will have its values modified. This commit ensures that the sliced Values are copied instead of returning the slice.